azure_function_app/requirements.txt
azure-functions
azure-storage-blob
requests
beautifulsoup4
lxml
python-dateutil
